Extracting verdict server code
Create a verdict.go that judges the output of a code execution, returning the status string "AC", "WA", "TLE" or "RTE"
